# docs/swagger.yaml
basePath: /api
definitions:
  dto.AddAccident:
    properties:
      accident_date:
        type: string
      accident_time:
        type: string
      accident_type:
        type: string
      cause_of_accident:
        type: string
      city_id:
        type: string
      city_name:
        type: string
      death_count:
        type: integer
      description:
        type: string
      district_id:
        type: string
      district_name:
        type: string
      injured_count:
        type: integer
      latitude:
        type: number
      location:
        maxLength: 255
        minLength: 5
        type: string
      longitude:
        type: number
      minor_injured_count:
        type: integer
      officer_name:
        type: string
      police_report_no:
        type: string
      police_station:
        type: string
      province_id:
        type: string
      province_name:
        type: string
      road_condition:
        type: string
      road_type:
        type: string
      vehicle_count:
        type: integer
      vehicle_type:
        type: string
      weather_condition:
        type: string
    required:
    - accident_date
    - accident_time
    - accident_type
    - city_id
    - city_name
    - district_id
    - district_name
    - location
    - police_report_no
    - province_id
    - province_name
    - vehicle_type
    type: object
...
  dto.UpdateAccident:
    properties:
      accident_date:
        type: string
      accident_time:
        type: string
      accident_type:
        type: string
      cause_of_accident:
        type: string
      city_id:
        type: string
      city_name:
        type: string
      death_count:
        type: integer
      description:
        type: string
      district_id:
        type: string
      district_name:
        type: string
      injured_count:
        type: integer
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      minor_injured_count:
        type: integer
      officer_name:
        type: string
      police_report_no:
        type: string
      police_station:
        type: string
      province_id:
        type: string
      province_name:
        type: string
      road_condition:
        type: string
      road_type:
        type: string
      vehicle_count:
        type: integer
      vehicle_type:
        type: string
      weather_condition:
        type: string
    type: object
...
  response.Error:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: zaiduszhuhur@gmail.com
    name: API Support
    url: https://www.linkedin.com/in/zaidus-zhuhur/
  description: This is a sample server for a safety riding service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Safety Riding API
  version: "1.0"
paths:
  /accident:
    post:
      consumes:
      - application/json
      description: Create a new accident record complete with location, reporter,
        and chronology information
      parameters:
      - description: Accident payload
        in: body
        name: accident
        required: true
        schema:
          $ref: '#/definitions/dto.AddAccident'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a new accident record
      tags:
      - Accidents
  /accident/{id}:
    delete:
      consumes:
      - application/json
      description: Delete accident by ID
      parameters:
      - description: Accident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete an accident
      tags:
      - Accidents
    get:
      consumes:
      - application/json
      description: Retrieve accident detail information by accident ID
      parameters:
      - description: Accident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Get accident details
      tags:
      - Accidents
    put:
      consumes:
      - application/json
      description: Update accident record fields by accident ID
      parameters:
      - description: Accident ID
        in: path
        name: id
        required: true
        type: string
      - description: Accident payload
        in: body
        name: accident
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccident'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Update accident data
      tags:
      - Accidents
  /accident/{id}/photos:
    post:
      consumes:
      - multipart/form-data
      description: Upload one or more accident photos and metadata
      parameters:
      - description: Accident ID
        in: path
        name: id
        required: true
        type: string
      - description: Accident photos
        in: formData
        name: photos
        required: true
        type: file
      - description: Photo captions (repeat per photo)
        in: formData
        name: captions
        type: string
      - description: Photo order (repeat per photo)
        in: formData
        name: photo_orders
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Upload accident photos
      tags:
      - Accidents
  /accident/photo/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete a single accident photo by photo ID
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete accident photo
      tags:
      - Accidents
  /accidents:
    get:
      consumes:
      - application/json
      description: Get paginated accidents with optional filters (district, city,
        province, type, etc.)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - description: Filter by district ID
        in: query
        name: district_id
        type: string
      - description: Filter by city ID
        in: query
        name: city_id
        type: string
      - description: Filter by province ID
        in: query
        name: province_id
        type: string
      - description: Filter by accident type
        in: query
        name: accident_type
        type: string
      - description: Filter by vehicle type
        in: query
        name: vehicle_type
        type: string
      - description: Filter by police station
        in: query
        name: police_station
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: List accidents with pagination
      tags:
      - Accidents
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"