# internal/handlers/http/accident/accident.go
package handleraccident

import (
	"errors"
	"fmt"
	"net/http"
	"reflect"
	"safety-riding/internal/dto"
	interfaceaccident "safety-riding/internal/interfaces/accident"
	"safety-riding/pkg/filter"
	"safety-riding/pkg/logger"
	"safety-riding/pkg/messages"
	"safety-riding/pkg/response"
	"safety-riding/utils"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type AccidentHandler struct {
	Service interfaceaccident.ServiceAccidentInterface
}

func NewAccidentHandler(s interfaceaccident.ServiceAccidentInterface) *AccidentHandler {
	return &AccidentHandler{
		Service: s,
	}
}

// AddAccident godoc
// @Summary Create a new accident record
// @Description Create a new accident record complete with location, reporter, and chronology information
// @Tags Accidents
// @Accept json
// @Produce json
// @Param accident body dto.AddAccident true "Accident payload"
// @Success 201 {object} response.Success
// @Failure 400 {object} response.Error
// @Failure 500 {object} response.Error
// @Security ApiKeyAuth
// @Router /accident [post]
func (h *AccidentHandler) AddAccident(ctx *gin.Context) {
	authData := utils.GetAuthData(ctx)
	username := utils.InterfaceString(authData["username"])
	logId := utils.GenerateLogId(ctx)
	logPrefix := fmt.Sprintf("[%s][AccidentHandler][AddAccident]", logId)

	var req dto.AddAccident
	if err := ctx.BindJSON(&req); err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; BindJSON ERROR: %s;", logPrefix, err.Error()))
		res := response.Response(http.StatusBadRequest, messages.InvalidRequest, logId, nil)
		res.Error = utils.ValidateError(err, reflect.TypeOf(req), "json")
		ctx.JSON(http.StatusBadRequest, res)
		return
	}
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Request: %+v;", logPrefix, utils.JsonEncode(req)))

	data, err := h.Service.AddAccident(username, req)
	if err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Service.AddAccident; Error: %+v", logPrefix, err))
		res := response.Response(http.StatusInternalServerError, messages.MsgFail, logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusInternalServerError, res)
		return
	}

	res := response.Response(http.StatusCreated, "Add accident successfully", logId, data)
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Success: %+v;", logPrefix, utils.JsonEncode(data)))
	ctx.JSON(http.StatusCreated, res)
}

// GetAccidentById godoc
// @Summary Get accident details
// @Description Retrieve accident detail information by accident ID
// @Tags Accidents
// @Accept json
// @Produce json
// @Param id path string true "Accident ID"
// @Success 200 {object} response.Success
// @Failure 404 {object} response.Error
// @Failure 500 {object} response.Error
// @Security ApiKeyAuth
// @Router /accident/{id} [get]
func (h *AccidentHandler) GetAccidentById(ctx *gin.Context) {
	logId := utils.GenerateLogId(ctx)
	logPrefix := fmt.Sprintf("[%s][AccidentHandler][GetAccidentById]", logId)

	accidentId, err := utils.ValidateUUID(ctx, logId)
	if err != nil {
		return
	}

	data, err := h.Service.GetAccidentById(accidentId)
	if err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Service.GetAccidentById; Error: %+v", logPrefix, err))
		if errors.Is(err, gorm.ErrRecordNotFound) {
			res := response.Response(http.StatusNotFound, messages.NotFound, logId, nil)
			res.Error = "accident data not found"
			ctx.JSON(http.StatusNotFound, res)
			return
		}

		res := response.Response(http.StatusInternalServerError, messages.MsgFail, logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusInternalServerError, res)
		return
	}

	res := response.Response(http.StatusOK, "Get accident successfully", logId, data)
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Response: %+v", logId, nil))
	ctx.JSON(http.StatusOK, res)
}

// UpdateAccident godoc
// @Summary Update accident data
// @Description Update accident record fields by accident ID
// @Tags Accidents
// @Accept json
// @Produce json
// @Param id path string true "Accident ID"
// @Param accident body dto.UpdateAccident true "Accident payload"
// @Success 200 {object} response.Success
// @Failure 400 {object} response.Error
// @Failure 500 {object} response.Error
// @Security ApiKeyAuth
// @Router /accident/{id} [put]
func (h *AccidentHandler) UpdateAccident(ctx *gin.Context) {
	authData := utils.GetAuthData(ctx)
	username := utils.InterfaceString(authData["username"])
	logId := utils.GenerateLogId(ctx)
	logPrefix := fmt.Sprintf("[%s][AccidentHandler][UpdateAccident]", logId)

	accidentId, err := utils.ValidateUUID(ctx, logId)
	if err != nil {
		return
	}

	var req dto.UpdateAccident
	if err := ctx.BindJSON(&req); err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; BindJSON ERROR: %s;", logPrefix, err.Error()))
		res := response.Response(http.StatusBadRequest, messages.InvalidRequest, logId, nil)
		res.Error = utils.ValidateError(err, reflect.TypeOf(req), "json")
		ctx.JSON(http.StatusBadRequest, res)
		return
	}
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Request: %+v;", logPrefix, utils.JsonEncode(req)))

	data, err := h.Service.UpdateAccident(accidentId, username, req)
	if err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Service.UpdateAccident; Error: %+v", logPrefix, err))
		res := response.Response(http.StatusInternalServerError, messages.MsgFail, logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusInternalServerError, res)
		return
	}

	res := response.Response(http.StatusOK, "Update accident successfully", logId, data)
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Success: %+v;", logPrefix, utils.JsonEncode(data)))
	ctx.JSON(http.StatusOK, res)
}

// FetchAccident godoc
// @Summary List accidents with pagination
// @Description Get paginated accidents with optional filters (district, city, province, type, etc.)
// @Tags Accidents
// @Accept json
// @Produce json
// @Param page query int false "Page number"
// @Param limit query int false "Items per page"
// @Param sort query string false "Sort field"
// @Param order query string false "Sort order (asc/desc)"
// @Param district_id query string false "Filter by district ID"
// @Param city_id query string false "Filter by city ID"
// @Param province_id query string false "Filter by province ID"
// @Param accident_type query string false "Filter by accident type"
// @Param vehicle_type query string false "Filter by vehicle type"
// @Param police_station query string false "Filter by police station"
// @Success 200 {object} response.Success
// @Failure 404 {object} response.Error
// @Failure 500 {object} response.Error
// @Security ApiKeyAuth
// @Router /accidents [get]
func (h *AccidentHandler) FetchAccident(ctx *gin.Context) {
	logId := utils.GenerateLogId(ctx)
	logPrefix := fmt.Sprintf("[%s][AccidentHandler][FetchAccident]", logId)

	params, _ := filter.GetBaseParams(ctx, "accident_date", "desc", 10)
	params.Filters = filter.WhitelistStringFilter(params.Filters, []string{"district_id", "city_id", "province_id", "accident_type", "vehicle_type", "police_station"})

	accidents, totalData, err := h.Service.FetchAccident(params)
	if err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Fetch; Error: %+v", logPrefix, err))
		if errors.Is(err, gorm.ErrRecordNotFound) {
			res := response.Response(http.StatusNotFound, messages.NotFound, logId, nil)
			res.Error = "List accident not found"
			ctx.JSON(http.StatusNotFound, res)
			return
		}

		res := response.Response(http.StatusInternalServerError, messages.MsgFail, logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusInternalServerError, res)
		return
	}

	res := response.PaginationResponse(http.StatusOK, int(totalData), params.Page, params.Limit, logId, accidents)
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Response: %+v;", logPrefix, utils.JsonEncode(accidents)))
	ctx.JSON(http.StatusOK, res)
}

// DeleteAccident godoc
// @Summary Delete an accident
// @Description Delete accident by ID
// @Tags Accidents
// @Accept json
// @Produce json
// @Param id path string true "Accident ID"
// @Success 200 {object} response.Success
// @Failure 400 {object} response.Error
// @Failure 500 {object} response.Error
// @Security ApiKeyAuth
// @Router /accident/{id} [delete]
func (h *AccidentHandler) DeleteAccident(ctx *gin.Context) {
	authData := utils.GetAuthData(ctx)
	username := utils.InterfaceString(authData["username"])
	logId := utils.GenerateLogId(ctx)
	logPrefix := fmt.Sprintf("[%s][AccidentHandler][DeleteAccident]", logId)

	id := ctx.Param("id")
	if id == "" {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Missing accident ID", logPrefix))
		res := response.Response(http.StatusBadRequest, "Accident ID is required", logId, nil)
		ctx.JSON(http.StatusBadRequest, res)
		return
	}

	if err := h.Service.DeleteAccident(id, username); err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Service.DeleteAccident; Error: %+v", logPrefix, err))
		res := response.Response(http.StatusInternalServerError, messages.MsgFail, logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusInternalServerError, res)
		return
	}

	res := response.Response(http.StatusOK, "Delete accident successfully", logId, nil)
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Success;", logPrefix))
	ctx.JSON(http.StatusOK, res)
}

// AddAccidentPhotos godoc
// @Summary Upload accident photos
// @Description Upload one or more accident photos and metadata
// @Tags Accidents
// @Accept multipart/form-data
// @Produce json
// @Param id path string true "Accident ID"
// @Param photos formData file true "Accident photos"
// @Param captions formData string false "Photo captions (repeat per photo)"
// @Param photo_orders formData int false "Photo order (repeat per photo)"
// @Success 201 {object} response.Success
// @Failure 400 {object} response.Error
// @Failure 500 {object} response.Error
// @Security ApiKeyAuth
// @Router /accident/{id}/photos [post]
func (h *AccidentHandler) AddAccidentPhotos(ctx *gin.Context) {
	authData := utils.GetAuthData(ctx)
	username := utils.InterfaceString(authData["username"])
	logId := utils.GenerateLogId(ctx)
	logPrefix := fmt.Sprintf("[%s][AccidentHandler][AddAccidentPhotos]", logId)

	accidentId := ctx.Param("id")
	if accidentId == "" {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Missing accident ID", logPrefix))
		res := response.Response(http.StatusBadRequest, "Accident ID is required", logId, nil)
		ctx.JSON(http.StatusBadRequest, res)
		return
	}

	// Parse multipart form
	form, err := ctx.MultipartForm()
	if err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; MultipartForm ERROR: %s;", logPrefix, err.Error()))
		res := response.Response(http.StatusBadRequest, "Failed to parse form data", logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusBadRequest, res)
		return
	}

	// Get files from form
	files := form.File["photos"]
	if len(files) == 0 {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; No photos uploaded", logPrefix))
		res := response.Response(http.StatusBadRequest, "At least one photo is required", logId, nil)
		ctx.JSON(http.StatusBadRequest, res)
		return
	}

	// Get optional captions and photo orders from form data
	captions := form.Value["captions"]
	photoOrders := form.Value["photo_orders"]

	// Call service to upload photos to MinIO and save to database
	data, err := h.Service.AddAccidentPhotosFromFiles(ctx, accidentId, username, files, captions, photoOrders)
	if err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Service.AddAccidentPhotosFromFiles; Error: %+v", logPrefix, err))
		res := response.Response(http.StatusInternalServerError, messages.MsgFail, logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusInternalServerError, res)
		return
	}

	res := response.Response(http.StatusCreated, "Add accident photos successfully", logId, data)
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Success: uploaded %d photos;", logPrefix, len(files)))
	ctx.JSON(http.StatusCreated, res)
}

// DeleteAccidentPhoto godoc
// @Summary Delete accident photo
// @Description Delete a single accident photo by photo ID
// @Tags Accidents
// @Accept json
// @Produce json
// @Param photoId path string true "Photo ID"
// @Success 200 {object} response.Success
// @Failure 400 {object} response.Error
// @Failure 500 {object} response.Error
// @Security ApiKeyAuth
// @Router /accident/photo/{photoId} [delete]
func (h *AccidentHandler) DeleteAccidentPhoto(ctx *gin.Context) {
	authData := utils.GetAuthData(ctx)
	username := utils.InterfaceString(authData["username"])
	logId := utils.GenerateLogId(ctx)
	logPrefix := fmt.Sprintf("[%s][AccidentHandler][DeleteAccidentPhoto]", logId)

	photoId := ctx.Param("photoId")
	if photoId == "" {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Missing photo ID", logPrefix))
		res := response.Response(http.StatusBadRequest, "Photo ID is required", logId, nil)
		ctx.JSON(http.StatusBadRequest, res)
		return
	}

	if err := h.Service.DeleteAccidentPhoto(photoId, username); err != nil {
		logger.WriteLog(logger.LogLevelError, fmt.Sprintf("%s; Service.DeleteAccidentPhoto; Error: %+v", logPrefix, err))
		res := response.Response(http.StatusInternalServerError, messages.MsgFail, logId, nil)
		res.Error = err.Error()
		ctx.JSON(http.StatusInternalServerError, res)
		return
	}

	res := response.Response(http.StatusOK, "Delete accident photo successfully", logId, nil)
	logger.WriteLog(logger.LogLevelDebug, fmt.Sprintf("%s; Success;", logPrefix))
	ctx.JSON(http.StatusOK, res)
}